+++ CHANGELOG.md
@@ -6,10 +6,13 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
-## [0.8.0] - 2023-08-24
-
 ### Added
 - `Trim`, `TrimStart`, `TrimEnd` are added to the string builder
+
+### Fixed
+- Sometimes the grow mechanism did not work properly leading to a panic
+
+## [0.8.0] - 2023-08-24
 
 ### Added